[TIMOB-23671] Major rearchitecture that cleans up the hacky old design, fixes issues with deadlocks, runloops, and debug logging. log() and installApp() are still not working.
