+++ src/runloop.h
@@ -1,0 +1,18 @@
+/**
+ * node-ios-device
+ * Copyright (c) 2013-2016 by Appcelerator, Inc. All Rights Reserved.
+ * Licensed under the terms of the Apache Public License
+ * Please see the LICENSE included with this distribution for details.
+ */
+
+#ifndef __RUNLOOP_H__
+#define __RUNLOOP_H__
+
+namespace node_ios_device {
+
+void startRunLoop();
+void stopRunLoop();
+
+} // end namespace node_ios_device
+
+#endif
